+++ elements/pl-symbolic-input/pl-symbolic-input.py
@@ -19,7 +19,7 @@ def get_variables_list(variables_string):
 def prepare(element_html, data):
     element = lxml.html.fragment_fromstring(element_html)
     required_attribs = ['answers-name']
-    optional_attribs = ['weight', 'correct-answer', 'variables', 'label', 'display']
+    optional_attribs = ['weight', 'correct-answer', 'variables', 'label', 'display', 'allow-complex', 'imaginary-unit-for-display']
     pl.check_attribs(element, required_attribs, optional_attribs)
     name = pl.get_string_attrib(element, 'answers-name')
 
@@ -29,6 +29,10 @@ def prepare(element_html, data):
             raise Exception('duplicate correct-answers variable name: %s' % name)
         data['correct-answers'][name] = correct_answer
 
+    imaginary_unit = pl.get_string_attrib(element, 'imaginary-unit-for-display', 'i')
+    if not (imaginary_unit == 'i' or imaginary_unit == 'j'):
+        raise Exception('imaginary-unit-for-display must be either i or j')
+
 
 def render(element_html, data):
     element = lxml.html.fragment_fromstring(element_html)
@@ -37,15 +41,23 @@ def render(element_html, data):
     variables_string = pl.get_string_attrib(element, 'variables', None)
     variables = get_variables_list(variables_string)
     display = pl.get_string_attrib(element, 'display', 'inline')
+    allow_complex = pl.get_boolean_attrib(element, 'allow-complex', False)
+    imaginary_unit = pl.get_string_attrib(element, 'imaginary-unit-for-display', 'i')
 
     if data['panel'] == 'question':
         editable = data['editable']
         raw_submitted_answer = data['raw_submitted_answers'].get(name, None)
 
         operators = ', '.join(['cos', 'sin', 'tan', 'exp', 'log', 'sqrt', '( )', '+', '-', '*', '/', '^', '**'])
-        constants = ', '.join(['pi'])
+        constants = ', '.join(['pi, e'])
 
-        info_params = {'format': True, 'variables': variables_string, 'operators': operators, 'constants': constants}
+        info_params = {
+            'format': True,
+            'variables': variables_string,
+            'operators': operators,
+            'constants': constants,
+            'allow_complex': allow_complex,
+        }
         with open('pl-symbolic-input.mustache', 'r', encoding='utf-8') as f:
             info = chevron.render(f, info_params).strip()
         with open('pl-symbolic-input.mustache', 'r', encoding='utf-8') as f:
@@ -60,7 +72,8 @@ def render(element_html, data):
             'editable': editable,
             'info': info,
             'shortinfo': shortinfo,
-            'uuid': pl.get_uuid()
+            'uuid': pl.get_uuid(),
+            'allow_complex': allow_complex,
         }
 
         partial_score = data['partial_scores'].get(name, {'score': None})
@@ -98,7 +111,12 @@ def render(element_html, data):
         }
         if parse_error is None:
             a_sub = data['submitted_answers'][name]
-            a_sub = phs.convert_string_to_sympy(a_sub, variables)
+            if isinstance(a_sub, str):
+                # this is for backward-compatibility
+                a_sub = phs.convert_string_to_sympy(a_sub, variables, allow_complex=allow_complex)
+            else:
+                a_sub = phs.json_to_sympy(a_sub, allow_complex=allow_complex)
+            a_sub = a_sub.subs(sympy.I, sympy.Symbol(imaginary_unit))
             html_params['a_sub'] = sympy.latex(a_sub)
         else:
             raw_submitted_answer = data['raw_submitted_answers'].get(name, None)
@@ -130,10 +148,14 @@ def render(element_html, data):
             html = chevron.render(f, html_params).strip()
 
     elif data['panel'] == 'answer':
-        a_tru = pl.from_json(data['correct_answers'].get(name, None))
+        a_tru = data['correct_answers'].get(name, None)
         if a_tru is not None:
             if isinstance(a_tru, str):
-                a_tru = phs.convert_string_to_sympy(a_tru, variables)
+                # this is so instructors can specify the true answer simply as a string
+                a_tru = phs.convert_string_to_sympy(a_tru, variables, allow_complex=allow_complex)
+            else:
+                a_tru = phs.json_to_sympy(a_tru, allow_complex=allow_complex)
+            a_tru = a_tru.subs(sympy.I, sympy.Symbol(imaginary_unit))
             html_params = {
                 'answer': True,
                 'label': label,
@@ -154,6 +176,8 @@ def parse(element_html, data):
     element = lxml.html.fragment_fromstring(element_html)
     name = pl.get_string_attrib(element, 'answers-name')
     variables = get_variables_list(pl.get_string_attrib(element, 'variables', None))
+    allow_complex = pl.get_boolean_attrib(element, 'allow-complex', False)
+    imaginary_unit = pl.get_string_attrib(element, 'imaginary-unit-for-display', 'i')
 
     # Get submitted answer or return parse_error if it does not exist
     a_sub = data['submitted_answers'].get(name, None)
@@ -172,10 +196,18 @@ def parse(element_html, data):
         a_sub = a_sub.strip()
 
         # Convert safely to sympy
-        a_sub_parsed = phs.convert_string_to_sympy(a_sub, variables)
+        a_sub_parsed = phs.convert_string_to_sympy(a_sub, variables, allow_complex=allow_complex)
 
-        # Store result as a string.
-        a_sub_string = str(a_sub_parsed)
+        # If complex numbers are not allowed, raise error if expression has the imaginary unit
+        if (not allow_complex) and (a_sub_parsed.has(sympy.I)):
+            a_sub_parsed = a_sub_parsed.subs(sympy.I, sympy.Symbol(imaginary_unit))
+            s = 'Your answer was simplified to this, which contains a complex number (denoted ${:s}$): $${:s}$$'.format(imaginary_unit, sympy.latex(a_sub_parsed))
+            data['format_errors'][name] = s
+            data['submitted_answers'][name] = None
+            return
+
+        # Store result as json.
+        a_sub_json = phs.sympy_to_json(a_sub_parsed, allow_complex=allow_complex)
     except phs.HasFloatError as err:
         s = 'Your answer contains the floating-point number ' + str(err.n) + '. '
         s += 'All numbers must be expressed as integers (or ratios of integers). '
@@ -186,6 +218,8 @@ def parse(element_html, data):
     except phs.HasComplexError as err:
         s = 'Your answer contains the complex number ' + str(err.n) + '. '
         s += 'All numbers must be expressed as integers (or ratios of integers). '
+        if allow_complex:
+            s += 'To include a complex number in your expression, write it as the product of an integer with the imaginary unit <code>i</code> or <code>j</code>. '
         s += '<br><br><pre>' + phs.point_to_error(a_sub, err.offset) + '</pre>'
         data['format_errors'][name] = s
         data['submitted_answers'][name] = None
@@ -231,13 +265,13 @@ def parse(element_html, data):
         data['submitted_answers'][name] = None
         return
 
-    # Make sure we can parse the string again, with the same set of variables
+    # Make sure we can parse the json again
     try:
         # Convert safely to sympy
-        phs.convert_string_to_sympy(a_sub_string, variables)
+        phs.json_to_sympy(a_sub_json, allow_complex=allow_complex)
 
         # Finally, store the result
-        data['submitted_answers'][name] = a_sub_string
+        data['submitted_answers'][name] = a_sub_json
     except Exception as err:
         s = 'Your answer was simplified to this, which contains an invalid expression: $${:s}$$'.format(sympy.latex(a_sub_parsed))
         data['format_errors'][name] = s
@@ -247,13 +281,13 @@ def parse(element_html, data):
 def grade(element_html, data):
     element = lxml.html.fragment_fromstring(element_html)
     name = pl.get_string_attrib(element, 'answers-name')
-
-    # Get weight
+    variables = get_variables_list(pl.get_string_attrib(element, 'variables', None))
+    allow_complex = pl.get_boolean_attrib(element, 'allow-complex', False)
     weight = pl.get_integer_attrib(element, 'weight', 1)
 
     # Get true answer (if it does not exist, create no grade - leave it
     # up to the question code)
-    a_tru = pl.from_json(data['correct_answers'].get(name, None))
+    a_tru = data['correct_answers'].get(name, None)
     if a_tru is None:
         return
 
@@ -263,11 +297,19 @@ def grade(element_html, data):
         data['partial_scores'][name] = {'score': 0, 'weight': weight}
         return
 
-    # Parse both correct and submitted answer (will throw an error on fail).
-    variables = get_variables_list(pl.get_string_attrib(element, 'variables', None))
+    # Parse true answer
     if isinstance(a_tru, str):
-        a_tru = phs.convert_string_to_sympy(a_tru, variables)
-    a_sub = phs.convert_string_to_sympy(a_sub, variables)
+        # this is so instructors can specify the true answer simply as a string
+        a_tru = phs.convert_string_to_sympy(a_tru, variables, allow_complex=allow_complex)
+    else:
+        a_tru = phs.json_to_sympy(a_tru, allow_complex=allow_complex)
+
+    # Parse submitted answer
+    if isinstance(a_sub, str):
+        # this is for backward-compatibility
+        a_sub = phs.convert_string_to_sympy(a_sub, variables, allow_complex=allow_complex)
+    else:
+        a_sub = phs.json_to_sympy(a_sub, allow_complex=allow_complex)
 
     # Check equality
     correct = a_tru.equals(a_sub)