+++ ChangeLog.md
@@ -29,6 +29,8 @@
 
   * Add personal access token management (Nathan Walters).
 
+  * Add `allow-complex` attribute for `pl-symbolic-input` (Tim Bretl).
+
   * Fix `pl-file-editor` to allow display empty text editor and add option to include text from source file (Mariana Silva).
 
   * Fix HTML rendering by reverting `cheerio.js` to `0.22.0` (Matt West).
@@ -66,6 +68,8 @@
   * Fig bug in `pl-symbolic-input` to handle submission of function names without arguments (Tim Bretl).
 
   * Fix bug in `pl-symbolic-input` to handle submissions that simplify to invalid expressions (Tim Bretl).
+
+  * Fix bug in `pl-symbolic-input` to handle the sympy constants I and E properly (Tim Bretl).
 
   * Change `pl-code` to display code from a source file OR inline text (Mariana Silva).